requirements_base.txt
bleach==3.1.0
celery==4.2.1
coverage==4.5.1
deepmerge==0.0.4
flex==6.13.2                        # optional, manually installed dependency for drf-yasg
dj-database-url==0.5.0
django-cors-headers==2.1.0
django-filter==2.0.0
django-simple-history==2.4.0
django==2.0.13
django-extra-fields==1.0.0
django_mock_queries==2.1.1
djangorestframework-camel-case==1.0b1
djangorestframework-jsonapi==2.7.0
djangorestframework==3.9.2
djangorestframework_simplejwt==3.3
djoser==1.4.1
drf-yasg==1.12.1
gunicorn==19.9.0
mock==2.0.0
model-mommy==1.5.1
parameterized==0.6.1
pip-tools==3.3.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
redis==2.10.6
requests==2.20.0
responses==0.9.0
simplejson==3.16.0
social-auth-app-django==3.1.0
swagger_spec_validator==2.4.3       # optional, manually installed dependency for drf-yasg
whitenoise==3.3.1